refactor: extract shared aihc-hackage library from duplicated code
Consolidate Hackage/Stackage download, snapshot parsing, and cabal-file
utilities into a single shared library at tooling/aihc-hackage/.

Previously, nearly identical implementations existed in:
- components/aihc-parser/common/HackageSupport.hs (shell curl+tar)
- components/aihc-parser/app/stackage-progress/StackageProgress/Snapshot.hs
- components/aihc-parser-cli/src/Aihc/Parser/Bench/Tarball.hs (http-client)

The new library uses http-client (not shell commands) and exposes:
- Aihc.Hackage.Types: PackageSpec, formatPackage
- Aihc.Hackage.Cache: XDG cache layout
- Aihc.Hackage.Download: package downloading
- Aihc.Hackage.Stackage: snapshot fetching and parsing
- Aihc.Hackage.Cabal: condition evaluation, component file discovery
- Aihc.Hackage.Util: file-system utilities

The original consumers become thin adapters:
- HackageSupport.hs: converts strings to Syntax.ExtensionSetting/LanguageEdition
- Snapshot.hs: 3-function re-export layer
- Summary.hs: re-exports PackageSpec from aihc-hackage
- Tarball.hs: drops ~400 lines of duplicated code

Net reduction: ~660 lines removed across consumers.
